#11 - Fixing errors and implementing features:  - Extra day is now Day+0.5  - Fitness function now runs 30 simulations and calculates the average and standard deviation  - Selection operator now clones the solutions in the mating pool in order to ensure re-evaluation of the fitness  - Created a TermWrapper that is used to implement equal in Terms and remove duplicated entries

+++ madetools/core/src/main/java/com/velonuboso/made/core/common/api/IEventFactory.java
@@ -34,7 +34,7 @@ import javafx.scene.paint.Color;
 
 @ImplementedBy(targetClass = EventFactory.class, targetMode = ImplementedBy.Mode.SINGLETON)
 public interface IEventFactory {
-    public void setDay(int day);
+    public void setDay(float day);
     IEvent inhabitantExists(ICharacter inhabitant);
     IEvent worldExists(IWorld world);
     IEvent newDay();
